+++ MBKC_System/MBKC.Service/Services/Implementations/HangfireService.cs
@@ -308,11 +308,11 @@ namespace MBKC.Service.Services.Implementations
                             decimal finalToTalPriceSubstractDeliveryFee = order.FinalTotalPrice - order.DeliveryFee;
                             if (exchangeWallets.ContainsKey(store.StoreId))
                             {
-                                exchangeWallets[store.StoreId] += finalToTalPriceSubstractDeliveryFee - (finalToTalPriceSubstractDeliveryFee * order.Commission / 100);
+                                exchangeWallets[store.StoreId] += finalToTalPriceSubstractDeliveryFee - (finalToTalPriceSubstractDeliveryFee * (decimal)order.Commission / 100);
                             }
                             else
                             {
-                                exchangeWallets.Add(store.StoreId, finalToTalPriceSubstractDeliveryFee - (finalToTalPriceSubstractDeliveryFee * order.Commission / 100));
+                                exchangeWallets.Add(store.StoreId, finalToTalPriceSubstractDeliveryFee - (finalToTalPriceSubstractDeliveryFee * (decimal)order.Commission / 100));
                             }
                         }
                     }
